Update the logout logic
